singlehop/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
Fabric3==1.13.1.post1
idna==2.6
paramiko==2.4.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
yact==0.5.1
